Backend/requirements.txt
Flask
flask-cors
firebase-admin
openai
python-dotenv
google-cloud-firestore
paystackapi
requests
gunicorn 